Move OrderPopulator into the Spree::Order namespace.
